full with rendering of regex
