+++ testplan/web_ui/testing/src/AssertionPane/Assertion.js
@@ -24,6 +24,10 @@ import SummaryBaseAssertion from "./AssertionSummary";
 import AttachmentAssertion from "./AssertionTypes/AttachmentAssertions";
 import PlotlyAssertion from "./AssertionTypes/PlotlyAssertion";
 import AttachedDirAssertion from "./AssertionTypes/AttachedDirAssertion";
+import {
+  RegexAssertion,
+  RegexMatchLineAssertion,
+} from "./AssertionTypes/RegexAssertions";
 import { EXPAND_STATUS } from "../Common/defaults";
 
 /**
@@ -90,6 +94,12 @@ class Assertion extends Component {
       CodeLog: CodeLogAssertion,
       Plotly: PlotlyAssertion,
       Directory: AttachedDirAssertion,
+      RegexMatch: RegexAssertion,
+      RegexMatchNotExists: RegexAssertion,
+      RegexSearch: RegexAssertion,
+      RegexSearchNotExists: RegexAssertion,
+      RegexFindIter: RegexAssertion,
+      RegexMatchLine: RegexMatchLineAssertion,
     };
     if (assertionMap[assertionType]) {
       return assertionMap[assertionType];
