+++ testplan/web_ui/testing/src/AssertionPane/AssertionTypes/__tests__/BasicAssertion.test.js
@@ -372,125 +372,6 @@ function propsExceptionNotRaised() {
   };
 }
 
-function propsRegexMatch() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.145337+00:00",
-      description: "string pattern match",
-      pattern: "foo",
-      line_no: 196,
-      meta_type: "assertion",
-      passed: true,
-      match_indexes: [[0, 3]],
-      type: "RegexMatch",
-      utc_time: "2019-02-12T17:41:43.145332+00:00",
-      string: "foobar",
-    },
-  };
-}
-
-function propsRegexMatchNotExists() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.150042+00:00",
-      description: null,
-      pattern: "baz",
-      line_no: 217,
-      meta_type: "assertion",
-      passed: true,
-      match_indexes: [],
-      type: "RegexMatchNotExists",
-      utc_time: "2019-02-12T17:41:43.150037+00:00",
-      string: "foobar",
-    },
-  };
-}
-
-function propsRegexSearch() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.153301+00:00",
-      description: null,
-      pattern: "second",
-      line_no: 225,
-      meta_type: "assertion",
-      passed: true,
-      match_indexes: [[11, 17]],
-      type: "RegexSearch",
-      utc_time: "2019-02-12T17:41:43.153294+00:00",
-      string: "first line\nsecond line\nthird line",
-    },
-  };
-}
-
-function propsRegexSearchNotExists() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.154688+00:00",
-      description: "Passing search empty",
-      pattern: "foobar",
-      line_no: 230,
-      meta_type: "assertion",
-      passed: true,
-      match_indexes: [],
-      type: "RegexSearchNotExists",
-      utc_time: "2019-02-12T17:41:43.154681+00:00",
-      string: "first line\nsecond line\nthird line",
-    },
-  };
-}
-
-function propsRegexFindIter() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.157637+00:00",
-      description: null,
-      string: "foo foo foo bar bar foo bar",
-      pattern: "foo",
-      line_no: 243,
-      condition_match: true,
-      meta_type: "assertion",
-      condition: "<lambda>",
-      passed: true,
-      type: "RegexFindIter",
-      utc_time: "2019-02-12T17:41:43.157630+00:00",
-      match_indexes: [
-        [0, 3],
-        [4, 7],
-        [8, 11],
-        [20, 23],
-      ],
-    },
-  };
-}
-
-function propsRegexMatchLine() {
-  return {
-    assertion: {
-      category: "DEFAULT",
-      machine_time: "2019-02-12T17:41:43.160987+00:00",
-      description: null,
-      string: "first line\nsecond line\nthird line",
-      pattern: "\\w+ line$",
-      line_no: 257,
-      meta_type: "assertion",
-      passed: true,
-      type: "RegexMatchLine",
-      utc_time: "2019-02-12T17:41:43.160981+00:00",
-      match_indexes: [
-        [0, 0, 10],
-        [1, 0, 11],
-        [2, 0, 10],
-      ],
-    },
-  };
-}
-
 function propsXMLCheck() {
   return {
     assertion: {
@@ -690,42 +571,6 @@ describe("BasicAssertion", () => {
 
   it("shallow renders the ExceptionNotRaised assertion", () => {
     props = propsExceptionNotRaised();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexMatch assertion", () => {
-    props = propsRegexMatch();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexMatchNotExists assertion", () => {
-    props = propsRegexMatchNotExists();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexSearch assertion", () => {
-    props = propsRegexSearch();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexSearchNotExists assertion", () => {
-    props = propsRegexSearchNotExists();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexFindIter assertion", () => {
-    props = propsRegexFindIter();
-    shallowComponent = shallow(<BasicAssertion {...props} />);
-    expect(shallowComponent).toMatchSnapshot();
-  });
-
-  it("shallow renders the RegexMatchLine assertion", () => {
-    props = propsRegexMatchLine();
     shallowComponent = shallow(<BasicAssertion {...props} />);
     expect(shallowComponent).toMatchSnapshot();
   });
