+++ testplan/web_ui/testing/src/AssertionPane/AssertionTypes/basicAssertionUtils.js
@@ -379,106 +379,6 @@ function prepareExceptionContent(assertion, defaultContent) {
   };
 }
 
-/*
- * Prepare the content for the RegexMatch, RegexMatchNotExists, RegexSearch,
- * RegexSearchNotExists and RegexFindIter assertions.
- * @param {object} assertion
- * @param {AssertionContent} defaultContent
- * @return {AssertionContent} Content for regex assertion
- * @private
- */
-function prepareRegexContent(assertion, defaultContent) {
-  const assertionString = assertion.string;
-  let reconstructedString = [];
-  let prevIdx = 0;
-  const uid = hashCode(JSON.stringify(assertion)).toString();
-
-  assertion.match_indexes.forEach((index) => {
-    reconstructedString.push(
-      <span key={uid + index + "0"}>
-        {assertionString.slice(prevIdx, index[0])}
-      </span>
-    );
-    reconstructedString.push(
-      <span
-        style={{ backgroundColor: "rgba(0, 123, 255, .5)" }}
-        key={uid + index + "1"}
-      >
-        {assertionString.slice(index[0], index[1])}
-      </span>
-    );
-    prevIdx = index[1];
-  });
-
-  reconstructedString.push(
-    <span key={uid + prevIdx}>{assertionString.slice(prevIdx)}</span>
-  );
-
-  const leftContent = (
-    <div>
-      <span>{assertion.pattern}</span>
-      <span>
-        {assertion.condition && (
-          <div key={assertion.uid}>
-            <br />
-            <strong>Condition:</strong>
-            <br />
-            <span>{assertion.condition}</span>
-          </div>
-        )}
-      </span>
-    </div>
-  );
-  const rightContent = (
-    <span style={{ whiteSpace: "pre" }}>{reconstructedString}</span>
-  );
-  const leftTitle = <span>Pattern:</span>;
-  const rightTitle = <span>String:</span>;
-
-  return {
-    ...defaultContent,
-    leftContent: leftContent,
-    rightContent: rightContent,
-    leftTitle: leftTitle,
-    rightTitle: rightTitle,
-  };
-}
-
-/**
- * Prepare the content for the RegexMatchLine assertion.
- * @param {object} assertion
- * @param {AssertionContent} defaultContent
- * @return {AssertionContent} Content for RegexMatchLine assertion
- * @private
- */
-function prepareRegexMatchLineContent(assertion, defaultContent) {
-  const assertionString = assertion.string.split("\n");
-  const uid = hashCode(JSON.stringify(assertion));
-  const reconstructedString = assertion.match_indexes.map((index) => (
-    <span
-      key={uid + index}
-      style={{ backgroundColor: "rgba(0, 123, 255, .5)" }}
-    >
-      {assertionString[index[0]].slice(index[1], index[2]) + "\n"}
-    </span>
-  ));
-
-  const leftContent = <span>{assertion.pattern}</span>;
-  const rightContent = (
-    <span style={{ whiteSpace: "pre" }}>{reconstructedString}</span>
-  );
-  const leftTitle = <span>Pattern:</span>;
-  const rightTitle = <span>String:</span>;
-
-  return {
-    ...defaultContent,
-    leftContent: leftContent,
-    rightContent: rightContent,
-    leftTitle: leftTitle,
-    rightTitle: rightTitle,
-  };
-}
-
 /**
  * Prepare the content for the XMLCheck assertion.
  * @param {object} assertion
@@ -726,16 +626,6 @@ function prepareBasicContent(assertion) {
     case "ExceptionRaised":
     case "ExceptionNotRaised":
       return prepareExceptionContent(assertion, defaultContent);
-
-    case "RegexMatch":
-    case "RegexMatchNotExists":
-    case "RegexSearch":
-    case "RegexSearchNotExists":
-    case "RegexFindIter":
-      return prepareRegexContent(assertion, defaultContent);
-
-    case "RegexMatchLine":
-      return prepareRegexMatchLineContent(assertion, defaultContent);
 
     case "XMLCheck":
       return prepareXMLCheckContent(assertion, defaultContent);
